move slack models
