+++ busy_beaver/models/__init__.py
@@ -4,7 +4,6 @@ from busy_beaver.common.models import BaseModel  # noqa isort:skip
 # Models
 from .event import Event  # noqa
 from .key_value import KeyValueStore  # noqa
-from .slack import SlackAppHomeOpened, SlackInstallation  # noqa
 from .task import (  # noqa
     PostGitHubSummaryTask,
     PostTweetTask,
@@ -15,4 +14,8 @@ from .youtube_video import YouTubeVideo  # noqa
 from busy_beaver.apps.github_integration.models import (  # noqa
     GitHubSummaryConfiguration,
     GitHubSummaryUser,
+)
+from busy_beaver.apps.slack_integration.models import (  # noqa
+    SlackAppHomeOpened,
+    SlackInstallation,
 )
